pyproject.toml
[project]
name = "batch-doc-vqa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ioannis Michaloliakos", email = "ioannis.michalol@ufl.edu" }
]
requires-python = ">=3.12"
dependencies = [
    "outlines>=1.2.3",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "rich>=14.1.0",
    "transformers",
    "requests",
]

[project.scripts]
batch-doc-vqa = "batch_doc_vqa:main"
openrouter-inference = "batch_doc_vqa.openrouter.main:main"
openrouter-cli = "batch_doc_vqa.openrouter.cli:main"
openrouter-ui = "batch_doc_vqa.openrouter.ui:interactive_organization_model_selection"
gemini-inference = "batch_doc_vqa.gemini.cli:main"
gemini-ui = "batch_doc_vqa.gemini.main:main"
update-benchmarks = "batch_doc_vqa.update_benchmarks:main"
generate-benchmark-table = "batch_doc_vqa.benchmarks.table_generator:main"
generate-pareto-plot = "batch_doc_vqa.benchmarks.pareto_plot:main"
update-readme-section = "batch_doc_vqa.benchmarks.readme_updater:main"

[project.optional-dependencies]
gemini = [
    "google-genai>=1.30.0",
]

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers.git" }

[dependency-groups]
dev = [
    "dotenv>=0.9.9",
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.5",
    "mypy>=1.17.1",
    "pandas>=2.3.1",
]
